Refactored the grandpa
